requirements.txt
Django==2.2.17
django-oauth-toolkit==1.0.0
django-cors-middleware==1.5.0
social-auth-core==2.0.0
social-auth-app-django==3.1.0
django-extensions==2.2.6
channels==2.4
channels_redis==3.2.0
django-extra-fields==0.9
pillow==8.0.1
#celery==4.2.1
gunicorn==20.0.4
urllib3<1.25,>=1.21.1
uvicorn[standard]==0.13.3
#daphne==2.2.2
pyyaml==5.3.1
watchdog==0.8.3
python-dateutil==2.7.3
bpython==0.17.1
websockets==8.1
aiofiles==0.4.0
oyaml==0.7
factory_boy==2.11.1
bleach==3.1.4
# Heroku staging debug tools
django-debug-toolbar==3.2
django-querycount==0.7.0

# User impersonation
django-su==0.9.0
django-ajax-selects==2.0.0

# Database
dj-database-url==0.4.2
psycopg2-binary==2.8.6
django-redis-cache==3.0.0
django_redis==4.12.1

# Storage
#-e git+https://github.com/jschneier/django-storages.git#egg=django-storages[azure,google]
django-storages[azure,google]==1.7.2
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
boto3==1.9.68
whitenoise==5.2.0

# Api
coreapi==2.3.3
djangorestframework==3.9.1
djangorestframework-csv==2.1.0
drf-extensions==0.4.0
markdown==2.6.11
pygments==2.2.0
drf-writable-nested==0.5.4
django-filter==2.4.0
drf-yasg[validation]==1.11.0
flex==6.12.0
#Click==7.0

# Greenlet/Gevent Workers
#gevent==1.4.0
#gevent-eventemitter==2.0
#greenlet==0.4.15

#urllib3<1.25,>=1.21.1
#tornado>=4.2.0,<6.0.0

pyrabbit2==1.0.7
django-enforce-host==1.0.1

# COS added
cos-python-sdk-v5==1.7.6 #Package qcloud_cos is imported by cos_storage

# pika
pika==1.2.0
# django-apscheduler apscheduler-3.7.0 django-apscheduler-0.5.2 tzlocal-2.1
django-apscheduler==0.5.2
